fix: アプリ全体で使われるComponentをcommonフォルダに移動・Loading関連の共通化
