+++ lib/ui/component/dialog/app_share_dialog.dart
@@ -6,7 +6,7 @@ import 'package:food_gram_app/core/model/posts.dart';
 import 'package:food_gram_app/core/model/users.dart';
 import 'package:food_gram_app/core/utils/helpers/share_helper.dart';
 import 'package:food_gram_app/gen/l10n/l10n.dart';
-import 'package:food_gram_app/ui/component/app_loading.dart';
+import 'package:food_gram_app/ui/component/common/app_loading.dart';
 import 'package:food_gram_app/ui/component/app_share_widget.dart';
 import 'package:gap/gap.dart';
 import 'package:go_router/go_router.dart';
@@ -230,7 +230,7 @@ class AppShareDialog extends HookConsumerWidget {
               ],
             ),
           ),
-          AppLoading(loading: loading.value, status: 'Loading...'),
+          AppProcessLoading(loading: loading.value, status: 'Loading...'),
         ],
       ),
     );
