+++ lib/ui/screen/edit_post/edit_post_screen.dart
@@ -11,7 +11,7 @@ import 'package:food_gram_app/core/utils/helpers/snack_bar_helper.dart';
 import 'package:food_gram_app/core/utils/provider/loading.dart';
 import 'package:food_gram_app/gen/l10n/l10n.dart';
 import 'package:food_gram_app/router/router.dart';
-import 'package:food_gram_app/ui/component/app_loading.dart';
+import 'package:food_gram_app/ui/component/common/app_loading.dart';
 import 'package:food_gram_app/ui/component/app_post_category_widget.dart';
 import 'package:food_gram_app/ui/component/app_text_field.dart';
 import 'package:food_gram_app/ui/component/modal_sheet/app_post_image_modal_sheet.dart';
@@ -251,7 +251,7 @@ class EditPostScreen extends HookConsumerWidget {
                 ),
               ),
             ),
-            AppLoading(
+            AppProcessLoading(
               loading: loading,
               status: state.status,
             ),
