+++ lib/ui/screen/paywall/paywall_screen.dart
@@ -9,7 +9,7 @@ import 'package:food_gram_app/core/theme/style/paywall_style.dart';
 import 'package:food_gram_app/core/utils/provider/loading.dart';
 import 'package:food_gram_app/gen/assets.gen.dart';
 import 'package:food_gram_app/gen/l10n/l10n.dart';
-import 'package:food_gram_app/ui/component/app_loading.dart';
+import 'package:food_gram_app/ui/component/common/app_loading.dart';
 import 'package:food_gram_app/ui/screen/setting/setting_view_model.dart';
 import 'package:gap/gap.dart';
 import 'package:go_router/go_router.dart';
@@ -163,7 +163,7 @@ class PaywallScreen extends ConsumerWidget {
               ],
             ),
           ),
-          AppLoading(
+          AppProcessLoading(
             loading: loading,
             status: 'Loading...',
           ),
