+++ lib/ui/screen/restaurant/restaurant_screen.dart
@@ -4,8 +4,8 @@ import 'package:food_gram_app/core/api/restaurant/repository/restaurant_reposito
 import 'package:food_gram_app/core/model/restaurant.dart';
 import 'package:food_gram_app/core/theme/style/restaurant_style.dart';
 import 'package:food_gram_app/gen/l10n/l10n.dart';
-import 'package:food_gram_app/ui/component/app_async_value_group.dart';
-import 'package:food_gram_app/ui/component/app_empty.dart';
+import 'package:food_gram_app/ui/component/common/app_async_value_group.dart';
+import 'package:food_gram_app/ui/component/common/app_empty.dart';
 import 'package:food_gram_app/ui/component/app_text_field.dart';
 import 'package:gap/gap.dart';
 import 'package:go_router/go_router.dart';
